Add types module and move AddQueueOption
